Move Asset class to Models directory
