+++ src/Transaction/TransactionFactory.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace pxgamer\Arionum\Transaction;
 
-use pxgamer\Arionum\Entity\Asset;
+use pxgamer\Arionum\Models\Asset;
 use pxgamer\Arionum\Models\Transaction;
 
 final class TransactionFactory
